fix : 좋아요 로직 수정 - CustomUserLikePostRepositoryImpl 에서 checkUserLikeStatusForPost 메서드 From 절 없던 구문 수정 - 이미 좋아요를 했는지 확인하는 로직 비즈니스 로직으로 분리 - 좋아요를 하는 로직이면 결과값에 true반환 - 좋아요를 한번 더 눌러서 삭제하는 로직이면 결과값에 false 반환

+++ src/main/java/com/usw/sugo/domain/majoruser/userlikepost/controller/UserLikePostController.java
@@ -9,6 +9,7 @@ import com.usw.sugo.domain.majoruser.UserLikePost;
 import com.usw.sugo.domain.majoruser.user.repository.UserRepository;
 import com.usw.sugo.domain.majoruser.userlikepost.dto.UserLikePostRequestDto.LikePostRequest;
 import com.usw.sugo.domain.majoruser.userlikepost.repository.UserLikePostRepository;
+import com.usw.sugo.domain.majoruser.userlikepost.service.UserLikePostService;
 import com.usw.sugo.global.exception.CustomException;
 import com.usw.sugo.global.exception.ErrorCode;
 import com.usw.sugo.global.jwt.JwtResolver;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 import java.util.HashMap;
+import java.util.Map;
 
 import static org.springframework.http.HttpStatus.OK;
 
@@ -30,7 +32,9 @@ public class UserLikePostController {
     private final ProductPostRepository productPostRepository;
 
     private final ProductPostFileRepository productPostFileRepository;
+    private final UserLikePostService userLikePostService;
     private final UserLikePostRepository userLikePostRepository;
+
     private final JwtResolver jwtResolver;
 
     /**
@@ -40,12 +44,9 @@ public class UserLikePostController {
      * @return
      */
     @PostMapping
-    public ResponseEntity<HashMap<String, Boolean>> likePost(
+    public ResponseEntity<Map<String, Boolean>> likePost(
             @RequestHeader String authorization,
             @RequestBody LikePostRequest likePostRequest) {
-        
-        // if 좋아요가 이미 박힌 게시글에 대한 요청이라면 좋아요 삭제
-
 
         User requestUser = userRepository.findById(
                 jwtResolver.jwtResolveToUserId(authorization.substring(7))).get();
@@ -58,7 +59,7 @@ public class UserLikePostController {
 
         
         // 좋아요를 하지 않은 게시글이면 좋아요 추가
-        if (userLikePostRepository.checkUserLikeStatusForPost(requestUser.getId(), productPost.getId())) {
+        if (!userLikePostService.isAlreadyLike(requestUser.getId(), productPost.getId())) {
 
             UserLikePost userLikePost = UserLikePost.builder()
                     .user(requestUser)
@@ -67,13 +68,21 @@ public class UserLikePostController {
                     .createdAt(LocalDateTime.now())
                     .build();
             userLikePostRepository.save(userLikePost);
+
+            return ResponseEntity
+                    .status(OK)
+                    .body(new HashMap<>() {{put("Like", true);
+                    }});
         }
 
         // 좋아요를 이미 했던 게시글이면 좋아요 삭제
-        userLikePostRepository.deleteLikePostByUserId(requestUser.getId(), likePostRequest.getProductPostId());
+        else if (userLikePostService.isAlreadyLike(requestUser.getId(), productPost.getId())) {
+            userLikePostRepository.deleteLikePostByUserId(requestUser.getId(), likePostRequest.getProductPostId());
+        }
 
-        return ResponseEntity.status(OK).body(new HashMap<>() {{
-            put("Success", true);
+        return ResponseEntity
+                .status(OK)
+                .body(new HashMap<>() {{put("Like", false);
         }});
     }
 }
